@@ -1,6 +1,6 @@
-import 'package:awesome_todo_app/data/database/data_source.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+import 'package:todo_shared_prefs/data/database/data_source.dart';
 
 import 'data/database/memory/memory_todo_repository.dart';
 import 'ui/list/todo_list.dart';
@@ -21,15 +21,15 @@ void main() async {
 }
 
 class AwesomeTodoApp extends StatelessWidget {
-  const AwesomeTodoApp({Key? key}) : super(key: key);
+  const AwesomeTodoApp({super.key});
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'AwesomeTodoApp',
       theme: ThemeData(
-        primarySwatch: Colors.blue,
-        visualDensity: VisualDensity.adaptivePlatformDensity,
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
       ),
       home: const TodoListPage(),
     );
