@@ -1,5 +1,5 @@
-import 'package:awesome_todo_app/data/database/data_source.dart';
-import 'package:awesome_todo_app/domain/model/todo.dart';
+import 'package:todo_shared_prefs/data/database/data_source.dart';
+import 'package:todo_shared_prefs/domain/model/todo.dart';
 import 'package:copy_with_extension/copy_with_extension.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
