@@ -1,17 +1,17 @@
 enum TodoPriority {
-  LOW,
-  NORMAL,
-  HIGH,
+  low,
+  normal,
+  high,
 }
 
 TodoPriority todoPriorityFromInt(int index) {
   switch (index) {
     case 0:
-      return TodoPriority.LOW;
+      return TodoPriority.low;
     case 1:
-      return TodoPriority.NORMAL;
+      return TodoPriority.normal;
     case 2:
-      return TodoPriority.HIGH;
+      return TodoPriority.high;
     default:
       throw ArgumentError(
           "Invalid Todo priority encountered while parsing from int");
